Extract inbox server helper modules
